+++ spec/slack/real_time/rtm_start_spec.rb
@@ -3,7 +3,7 @@
 
 #
 # to re-record a rtm_start fixture run with
-# SLACK_API_TOKEN=... CONCURRENCY=faye-websocket rspec spec/slack/real_time/rtm_start_spec.rb
+# SLACK_API_TOKEN=... CONCURRENCY=async-websocket rspec spec/slack/real_time/rtm_start_spec.rb
 # edit rtm_start.yml and remove the token, fix wss:// path (run specs, fix failures)
 #